+++ config/test.exs
@@ -1,6 +1,8 @@
 use Mix.Config
 
 config :autoformatbot,
-  github_owner: "maltoe",
-  github_repo: "autoformatbot.ex",
-  github_token: {:system, "GITHUB_TOKEN"}
+  adapter: {:github,
+    owner: "maltoe",
+    repo: "autoformatbot.ex",
+    access_token: {:system, "GITHUB_TOKEN"}
+  }
