+++ lib/autoformatbot/adapter.ex
@@ -1,0 +1,3 @@
+defmodule Autoformatbot.Adapter do
+  @callback 
+end
